Move everything related to `internals` into a separate detail header
