+++ setup.py
@@ -16,6 +16,7 @@ else:
         'include/pybind11/detail/common.h',
         'include/pybind11/detail/descr.h',
         'include/pybind11/detail/init.h',
+        'include/pybind11/detail/internals.h',
         'include/pybind11/detail/typeid.h'
         'include/pybind11/attr.h',
         'include/pybind11/buffer_info.h',
